Extract tray-window-renderer.js from renderer.js
