+++ tray-window.html
@@ -33,7 +33,7 @@
     margin-left: 147px;
 }
 </style>
-<script src="./renderer.js"></script>
+<script src="./tray-window-renderer.js"></script>
     </head>
     <body>
         <div class="container" name="timeform">
